+++ src/content-script/Root.tsx
@@ -8,7 +8,7 @@ import MainPage from '@/components/MainPage';
 import QueryBox from '@/components/QueryBox';
 import ThreadAnchor from '@/components/ThreadAnchor';
 import ThreadExportButton from '@/components/ThreadExportButton';
-import ThreadQueryFormatSwitch from '@/components/ThreadQueryFormatSwitch';
+import ThreadQueryHeader from '@/components/ThreadQueryHeader';
 import { Toaster } from '@/components/ui/toaster';
 import { useToast } from '@/components/ui/use-toast';
 import {
@@ -35,8 +35,12 @@ export default function Root() {
         <QueryBox />
         <Commander />
         {popupSettingsStore.getState().qolTweaks.threadTOC && <ThreadAnchor />}
-        {popupSettingsStore.getState().visualTweaks.threadQueryMarkdown && <ThreadQueryFormatSwitch />}
-        {popupSettingsStore.getState().qolTweaks.codeBlockEnhancedToolbar && <CodeBlockEnhancedToolbar />}
+        {popupSettingsStore.getState().visualTweaks.threadQueryEnhancedToolbar && (
+          <ThreadQueryHeader />
+        )}
+        {popupSettingsStore.getState().qolTweaks.codeBlockEnhancedToolbar && (
+          <CodeBlockEnhancedToolbar />
+        )}
         <ThreadExportButton />
         <Toaster />
         <IncompatibleInterfaceLanguageNotice />