+++ src/utils/observer.ts
@@ -288,6 +288,36 @@ function onShallowRouteChange(callback: () => void) {
   }).observe(document, { subtree: true, childList: true });
 }
 
+type ScrollCallback = () => void;
+
+function onScrollDirectionChange({
+  up,
+  down,
+  identifier,
+}: {
+  up?: ScrollCallback;
+  down?: ScrollCallback;
+  identifier: string;
+}) {
+  let lastScrollTop = 0;
+
+  $(window).on(`scroll.${identifier}`, function () {
+    const currentScrollTop = $(this).scrollTop();
+
+    if (typeof currentScrollTop === 'undefined') return;
+
+    if (currentScrollTop > lastScrollTop) {
+      down?.();
+    } else {
+      up?.();
+    }
+
+    lastScrollTop = currentScrollTop;
+  });
+
+  return () => $(window).off(`scroll.${identifier}`);
+}
+
 const observer = {
   onElementExist,
   onElementRemoved,
@@ -295,6 +325,7 @@ const observer = {
   onShallowRouteChange,
   onDOMChanges,
   onNewElementAdded,
+  onScrollDirectionChange,
 };
 
 export default observer;